moved filtering out of extract tests
