+++ pipelines/incidents-data/transform_incidents.py
@@ -1,6 +1,5 @@
 """A script to transform data from the National Rail Incidents API."""
 
-from os import environ as ENV
 import logging
 import re
